move group map related code to ec file
