+++ src/js_crypto/elliptic_curve.ts
@@ -1,5 +1,5 @@
-import { inverse, mod, p, q } from './finite_field.js';
-export { Pallas, Vesta, GroupAffine, GroupProjective };
+import { FiniteField, inverse, mod, p, q } from './finite_field.js';
+export { Pallas, Vesta, GroupAffine, GroupProjective, GroupMap };
 
 // TODO: constants, like generator points and cube roots for endomorphisms, should be drawn from
 // a common source, i.e. generated from the Rust code
@@ -29,6 +29,107 @@ const projectiveZero = { x: 1n, y: 1n, z: 0n };
 
 type GroupProjective = { x: bigint; y: bigint; z: bigint };
 type GroupAffine = { x: bigint; y: bigint; infinity: boolean };
+
+type GroupMapParams = {
+  u: bigint;
+  u_over_2: bigint;
+  conic_c: bigint;
+  projection_point: {
+    z: bigint;
+    y: bigint;
+  };
+  spec: { a: bigint; b: bigint };
+};
+
+type Conic = {
+  z: bigint;
+  y: bigint;
+};
+
+type STuple = {
+  u: bigint;
+  v: bigint;
+  y: bigint;
+};
+
+const GroupMap = {
+  Tock: (F: FiniteField) => {
+    const params: GroupMapParams = {
+      u: 2n,
+      u_over_2: 1n,
+      conic_c: 3n,
+      projection_point: {
+        z: 12196889842669319921865617096620076994180062626450149327690483414064673774441n,
+        y: 1n,
+      },
+      spec: {
+        a: 0n,
+        b: 5n,
+      },
+    };
+
+    function tryDecode(x: bigint): { x: bigint; y: bigint } | undefined {
+      const { a, b } = params.spec;
+
+      function f(x: bigint) {
+        // a * a * a = a^3
+        const pow3 = F.power(x, 3n);
+        // a * x
+        const ax = F.mul(a, x);
+        // a^3 + ax + b
+        return F.add(F.add(pow3, ax), b);
+      }
+
+      const y = f(x);
+      console.log(y);
+      const sqrtY = F.sqrt(y);
+      return F.isSquare(y) && sqrtY ? { x, y: sqrtY } : undefined;
+    }
+
+    function s_to_v_truncated(s: STuple): [bigint, bigint, bigint] {
+      const { u, v, y } = s;
+      return [v, F.negate(F.add(u, v)), F.add(u, F.square(y))];
+    }
+
+    function conic_to_s(c: Conic): STuple {
+      const d = F.div(c.z, c.y);
+      if (!d) throw Error(`Division undefined! ${c.z}/${c.y}`);
+      const v = F.sub(d, params.u_over_2);
+
+      return {
+        u: params.u,
+        v,
+        y: c.y,
+      };
+    }
+
+    function field_to_conic(t: bigint): Conic {
+      const { z: z0, y: y0 } = params.projection_point;
+
+      const ct = F.mul(params.conic_c, t);
+
+      const d1 = F.add(F.mul(ct, y0), z0);
+      const d2 = F.add(F.mul(ct, t), 1n);
+
+      const d = F.div(d1, d2);
+
+      if (!d) throw Error(`Division undefined! ${d1}/${d2}`);
+
+      const s = F.mul(2n, d);
+
+      return {
+        z: F.sub(z0, s),
+        y: F.sub(y0, F.mul(s, t)),
+      };
+    }
+
+    return {
+      potentialXs: (t: bigint) =>
+        s_to_v_truncated(conic_to_s(field_to_conic(t))),
+      tryDecode,
+    };
+  },
+};
 
 function projectiveNeg({ x, y, z }: GroupProjective, p: bigint) {
   return { x, y: y === 0n ? 0n : p - y, z };
